+++ src/providers/IvrApi.cpp
@@ -58,6 +58,30 @@ void IvrApi::getBulkEmoteSets(QString emoteSetList,
         .execute();
 }
 
+void IvrApi::getUserRoles(QString userName,
+                          ResultCallback<IvrResolve> successCallback,
+                          IvrFailureCallback failureCallback)
+{
+    assert(!userName.isEmpty());
+
+    this->makeRequest(QString("twitch/user"),
+                      QUrlQuery(QString("login=%1").arg(userName)))
+        .onSuccess([successCallback, failureCallback](auto result) -> Outcome {
+            auto root = result.parseJsonArray();
+
+            successCallback(root);
+
+            return Success;
+        })
+        .onError([failureCallback](auto result) {
+            qCWarning(chatterinoIvr)
+                << "Failed IVR API Call!" << result.status()
+                << QString(result.getData());
+            failureCallback();
+        })
+        .execute();
+}
+
 NetworkRequest IvrApi::makeRequest(QString url, QUrlQuery urlQuery)
 {
     assert(!url.startsWith("/"));