+++ src/widgets/splits/Split.cpp
@@ -106,15 +106,18 @@ Split::Split(QWidget *parent)
 
     this->input_->ui_.textEdit->installEventFilter(parent);
 
-    // update placeholder text on Twitch account change and channel change
-    this->bSignals_.emplace_back(
-        getApp()->accounts->twitch.currentUserChanged.connect([this] {
+    if (getSettings()->showTextInputPlaceholder)
+    {
+        // update placeholder text on Twitch account change and channel change
+        this->bSignals_.emplace_back(
+            getApp()->accounts->twitch.currentUserChanged.connect([this] {
+                this->updateInputPlaceholder();
+            }));
+        this->signalHolder_.managedConnect(channelChanged, [this] {
             this->updateInputPlaceholder();
-        }));
-    this->signalHolder_.managedConnect(channelChanged, [this] {
+        });
         this->updateInputPlaceholder();
-    });
-    this->updateInputPlaceholder();
+    }
 
     // clear SplitInput selection when selecting in ChannelView
     this->view_->selectionChanged.connect([this]() {
@@ -620,8 +623,9 @@ void Split::updateInputPlaceholder()
     else
     {
         placeholderText =
-            QString("Send message as %1...")
-                .arg(getApp()->accounts->twitch.getCurrent()->getUserName());
+            QString("Send message as %1 in %2...")
+                .arg(getApp()->accounts->twitch.getCurrent()->getUserName(),
+                     this->getChannel()->getName());
     }
 
     this->input_->ui_.textEdit->setPlaceholderText(placeholderText);